子弹穿透计算属性

+++ DungeonShooting_Godot/src/game/data/BulletData.cs
@@ -55,7 +55,7 @@ public class BulletData
     /// <summary>
     /// 子弹最大穿透次数
     /// </summary>
-    public float Penetration;
+    public int Penetration;
 
     /// <summary>
     /// 子弹最大存在时间
